Fix -Wsign-conversion and -Wconversion warnings (#71)

+++ color.h
@@ -40,7 +40,7 @@ typedef int (CSSAbortFunc)(void *);
 #define CSS_MEDIA_ALL     0xffff
 
 typedef uint32_t QEColor;
-#define QEARGB(a,r,g,b)    (((unsigned int)(a) << 24) | ((r) << 16) | ((g) << 8) | (b))
+#define QEARGB(a,r,g,b)    (((unsigned int)(a) << 24) | ((unsigned int)(r) << 16) | ((unsigned int)(g) << 8) | (unsigned int)(b))
 #define QERGB(r,g,b)       QEARGB(0xff, r, g, b)
 #define QERGB25(r,g,b)     QEARGB(1, r, g, b)
 #define COLOR_TRANSPARENT  0
@@ -195,7 +195,7 @@ int css_get_font_family(const char *str);
 int css_get_enum(const char *str, const char *enum_str);
 int color_dist(QEColor c1, QEColor c2);
 static inline int color_y(QEColor c) {
-    return 30 * ((c >> 16) & 0xff) + 59 * ((c >> 8) & 0xff) + 11 * (c & 0xff);
+    return (int)(30 * ((c >> 16) & 0xff) + 59 * ((c >> 8) & 0xff) + 11 * (c & 0xff));
 }
 
 typedef struct CSSRect {
