+++ cutils.h
@@ -175,9 +175,9 @@ static inline int strequal(const char *s1, const char *s2) {
  * p1 and p2 should have the same type
  * the macro checks that the types pointed to by p1 and p2 have the same size
  */
-#define blockcmp(p1, p2, n)   memcmp(p1, p2, (n) * sizeof *(p1) / (sizeof *(p1) == sizeof *(p2)))
-#define blockcpy(p1, p2, n)   memcpy(p1, p2, (n) * sizeof *(p1) / (sizeof *(p1) == sizeof *(p2)))
-#define blockmove(p1, p2, n) memmove(p1, p2, (n) * sizeof *(p1) / (sizeof *(p1) == sizeof *(p2)))
+#define blockcmp(p1, p2, n)   memcmp(p1, p2, (size_t)(n) * sizeof *(p1) / (sizeof *(p1) == sizeof *(p2)))
+#define blockcpy(p1, p2, n)   memcpy(p1, p2, (size_t)(n) * sizeof *(p1) / (sizeof *(p1) == sizeof *(p2)))
+#define blockmove(p1, p2, n) memmove(p1, p2, (size_t)(n) * sizeof *(p1) / (sizeof *(p1) == sizeof *(p2)))
 
 size_t get_basename_offset(const char *filename);
 
@@ -386,7 +386,7 @@ static inline int min3_int(int a, int b, int c) {
     return min_int(min_int(a, b), c);
 }
 
-static inline int max_uint(unsigned int a, unsigned int b) {
+static inline unsigned int max_uint(unsigned int a, unsigned int b) {
     /*@API utils
        Compute the maximum value of 2 integers
        @argument `a` an `unsigned int` value
@@ -399,7 +399,7 @@ static inline int max_uint(unsigned int a, unsigned int b) {
         return b;
 }
 
-static inline int min_uint(unsigned int a, unsigned int b) {
+static inline unsigned int min_uint(unsigned int a, unsigned int b) {
     /*@API utils
        Compute the minimum value of 2 integers
        @argument `a` an `unsigned int` value
